+++ src/rembrandt/__init__.py
@@ -2,6 +2,7 @@
 
 from importlib.metadata import version
 
+from rembrandt.conjugation import can_conjugate, conjugate
 from rembrandt.db import Database
 from rembrandt.lessons import lesson_progress, load_lessons
 from rembrandt.models import (
@@ -30,6 +31,8 @@ __all__ = [
     "Session",
     "SessionMode",
     "Word",
+    "can_conjugate",
+    "conjugate",
     "learning_mode",
     "lesson_progress",
     "load_lessons",